+++ crates/dns-resolver/src/lib.rs
@@ -1,4 +1,3 @@
-use anyhow::Context;
 use arc_swap::ArcSwap;
 use hickory_resolver::proto::op::ResponseCode;
 pub use hickory_resolver::proto::rr::rdata::tlsa::{CertUsage, Matching, Selector, TLSA};
@@ -9,17 +8,11 @@ use kumo_address::host::HostAddress;
 use kumo_address::host_or_socket::HostOrSocketAddress;
 use kumo_address::resolvable::ResolvableSocketAddr;
 use kumo_log_types::ResolvedAddress;
-use kumo_prometheus::declare_metric;
 use lruttl::declare_cache;
-use rand::prelude::SliceRandom;
 use serde::{Deserialize, Serialize};
-use std::collections::BTreeMap;
 use std::net::{IpAddr, Ipv6Addr, SocketAddr};
-use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::{Arc, LazyLock};
-use std::time::{Duration, Instant};
-use tokio::sync::Semaphore;
-use tokio::time::timeout;
+use std::time::Instant;
 
 mod resolver;
 #[cfg(feature = "unbound")]
@@ -37,10 +30,6 @@ static RESOLVER: LazyLock<ArcSwap<Box<dyn Resolver>>> =
     LazyLock::new(|| ArcSwap::from_pointee(Box::new(default_resolver())));
 
 declare_cache! {
-/// Caches domain name to computed set of MailExchanger records
-static MX_CACHE: LruCacheWithTtl<(Name, Option<u16>), Result<Arc<MailExchanger>, String>>::new("dns_resolver_mx", 64 * 1024);
-}
-declare_cache! {
 /// Caches domain name to ipv4 records and their DNSSEC secure status
 static IPV4_CACHE: LruCacheWithTtl<Name, Arc<IpAddresses>>::new("dns_resolver_ipv4", 1024);
 }
@@ -54,102 +43,11 @@ declare_cache! {
 static IP_CACHE: LruCacheWithTtl<(Name, IpLookupStrategy), Arc<IpAddresses>>::new("dns_resolver_ip", 1024);
 }
 
-/// Maximum number of concurrent mx resolves permitted
-static MX_MAX_CONCURRENCY: AtomicUsize = AtomicUsize::new(128);
-static MX_CONCURRENCY_SEMA: LazyLock<Semaphore> =
-    LazyLock::new(|| Semaphore::new(MX_MAX_CONCURRENCY.load(Ordering::SeqCst)));
-
-/// 5 seconds in ms
-static MX_TIMEOUT_MS: AtomicUsize = AtomicUsize::new(5000);
-
-/// 5 minutes in ms
-static MX_NEGATIVE_TTL: AtomicUsize = AtomicUsize::new(300 * 1000);
-
-declare_metric! {
-/// number of `MailExchanger::resolve` calls currently in progress.
-static MX_IN_PROGRESS: IntGauge("dns_mx_resolve_in_progress");
-}
-
-declare_metric! {
-/// Total number of successful `MailExchanger::resolve` calls
-static MX_SUCCESS: IntCounter(
-        "dns_mx_resolve_status_ok");
-}
-
-declare_metric! {
-/// Total number of failed `MailExchanger::resolve` calls.
-///
-/// Spikes may indicate an issue with your DNS configuration
-/// or infrastructure, or may simply indicate that the traffic
-/// is destined for bogus addresses.
-static MX_FAIL: IntCounter("dns_mx_resolve_status_fail");
-}
-
-declare_metric! {
-/// Total number of MailExchanger::resolve calls satisfied by level 1 cache.
-///
-/// Redundant with the newer [lruttl_hit_count{cache_name="dns_resolver_mx"}](lruttl_hit_count.md)
-/// metric.
-static MX_CACHED: IntCounter("dns_mx_resolve_cache_hit");
-}
-
-declare_metric! {
-/// Total number of MailExchanger::resolve calls that resulted in an MX DNS request to the next level of cache
-///
-/// Redundant with the newer [lruttl_miss_count{cache_name="dns_resolver_mx"}](lruttl_miss_count.md)
-/// metric.
-static MX_QUERIES: IntCounter("dns_mx_resolve_cache_miss");
-}
-
 fn default_resolver() -> impl Resolver {
     #[cfg(feature = "default-unbound")]
     return UnboundResolver::new().unwrap();
     #[cfg(not(feature = "default-unbound"))]
     return HickoryResolver::new().expect("Parsing /etc/resolv.conf failed");
-}
-
-pub fn set_mx_concurrency_limit(n: usize) {
-    MX_MAX_CONCURRENCY.store(n, Ordering::SeqCst);
-}
-
-pub fn set_mx_timeout(duration: Duration) -> anyhow::Result<()> {
-    let ms = duration
-        .as_millis()
-        .try_into()
-        .context("set_mx_timeout: duration is too large")?;
-    MX_TIMEOUT_MS.store(ms, Ordering::Relaxed);
-    Ok(())
-}
-
-pub fn get_mx_timeout() -> Duration {
-    Duration::from_millis(MX_TIMEOUT_MS.load(Ordering::Relaxed) as u64)
-}
-
-pub fn set_mx_negative_cache_ttl(duration: Duration) -> anyhow::Result<()> {
-    let ms = duration
-        .as_millis()
-        .try_into()
-        .context("set_mx_negative_cache_ttl: duration is too large")?;
-    MX_NEGATIVE_TTL.store(ms, Ordering::Relaxed);
-    Ok(())
-}
-
-pub fn get_mx_negative_ttl() -> Duration {
-    Duration::from_millis(MX_NEGATIVE_TTL.load(Ordering::Relaxed) as u64)
-}
-
-#[derive(Clone, Debug, Serialize)]
-pub struct MailExchanger {
-    pub domain_name: String,
-    pub hosts: Vec<String>,
-    pub site_name: String,
-    pub by_pref: BTreeMap<u16, Vec<String>>,
-    pub is_domain_literal: bool,
-    /// DNSSEC verified
-    pub is_secure: bool,
-    pub is_mx: bool,
-    #[serde(skip)]
-    expires: Option<Instant>,
 }
 
 pub fn fully_qualify(domain_name: &str) -> Result<Name, ProtoError> {
@@ -637,261 +535,6 @@ pub async fn resolve_socket_addr(
     }
 }
 
-impl MailExchanger {
-    pub async fn resolve(domain_name: &str) -> anyhow::Result<Arc<Self>> {
-        MX_IN_PROGRESS.inc();
-        let result = Self::resolve_impl(domain_name).await;
-        MX_IN_PROGRESS.dec();
-        if result.is_ok() {
-            MX_SUCCESS.inc();
-        } else {
-            MX_FAIL.inc();
-        }
-        result
-    }
-
-    async fn resolve_impl(domain_name: &str) -> anyhow::Result<Arc<Self>> {
-        let (name_fq, opt_port) = match DomainClassification::classify(domain_name)? {
-            DomainClassification::Literal(addr) => {
-                let mut by_pref = BTreeMap::new();
-                by_pref.insert(1, vec![addr.to_string()]);
-                return Ok(Arc::new(Self {
-                    domain_name: domain_name.to_string(),
-                    hosts: vec![addr.to_string()],
-                    site_name: addr.to_string(),
-                    by_pref,
-                    is_domain_literal: true,
-                    is_secure: false,
-                    is_mx: false,
-                    expires: None,
-                }));
-            }
-            DomainClassification::Domain(name_fq, opt_port) => (name_fq, opt_port),
-        };
-
-        let lookup_result = MX_CACHE
-            .get_or_try_insert(
-                &(name_fq.clone(), opt_port),
-                |mx_result| {
-                    if let Ok(mx) = mx_result {
-                        if let Some(exp) = mx.expires {
-                            return exp
-                                .checked_duration_since(std::time::Instant::now())
-                                .unwrap_or_else(|| Duration::from_secs(10));
-                        }
-                    }
-                    get_mx_negative_ttl()
-                },
-                async {
-                    MX_QUERIES.inc();
-                    let start = Instant::now();
-                    let (mut by_pref, expires) = match lookup_mx_record(&name_fq).await {
-                        Ok((by_pref, expires)) => (by_pref, expires),
-                        Err(err) => {
-                            let error = format!(
-                                "MX lookup for {domain_name} failed after {elapsed:?}: {err:#}",
-                                elapsed = start.elapsed()
-                            );
-                            return Ok::<Result<Arc<MailExchanger>, String>, anyhow::Error>(Err(
-                                error,
-                            ));
-                        }
-                    };
-
-                    let mut hosts = vec![];
-                    for pref in &mut by_pref {
-                        for host in &mut pref.hosts {
-                            if let Some(port) = opt_port {
-                                *host = format!("{host}:{port}");
-                            };
-                            hosts.push(host.to_string());
-                        }
-                    }
-
-                    let is_secure = by_pref.iter().all(|p| p.is_secure);
-                    let is_mx = by_pref.iter().all(|p| p.is_mx);
-
-                    let by_pref = by_pref
-                        .into_iter()
-                        .map(|pref| (pref.pref, pref.hosts))
-                        .collect();
-
-                    let site_name = factor_names(&hosts);
-                    let mx = Self {
-                        hosts,
-                        domain_name: name_fq.to_ascii(),
-                        site_name,
-                        by_pref,
-                        is_domain_literal: false,
-                        is_secure,
-                        is_mx,
-                        expires: Some(expires),
-                    };
-
-                    Ok(Ok(Arc::new(mx)))
-                },
-            )
-            .await
-            .map_err(|err| anyhow::anyhow!("{err}"))?;
-
-        if !lookup_result.is_fresh {
-            MX_CACHED.inc();
-        }
-
-        lookup_result.item.map_err(|err| anyhow::anyhow!("{err}"))
-    }
-
-    pub fn has_expired(&self) -> bool {
-        match self.expires {
-            Some(deadline) => deadline <= Instant::now(),
-            None => false,
-        }
-    }
-
-    /// Returns the list of resolve MX hosts in *reverse* preference
-    /// order; the first one to try is the last element.
-    /// smtp_dispatcher.rs relies on this ordering, as it will pop
-    /// off candidates until it has exhausted its connection plan.
-    pub async fn resolve_addresses(&self, strategy: IpLookupStrategy) -> ResolvedMxAddresses {
-        let mut result = vec![];
-
-        for hosts in self.by_pref.values().rev() {
-            let mut by_pref = vec![];
-
-            for mx_host in hosts {
-                // '.' is a null mx; skip trying to resolve it
-                if mx_host == "." {
-                    return ResolvedMxAddresses::NullMx;
-                }
-
-                // Handle the literal address case
-                let (mx_host, opt_port) = match has_colon_port(mx_host) {
-                    Some((domain_name, port)) => (domain_name, Some(port)),
-                    None => (mx_host.as_str(), None),
-                };
-                if let Ok(addr) = mx_host.parse::<IpAddr>() {
-                    let mut addr: HostOrSocketAddress = addr.into();
-                    if let Some(port) = opt_port {
-                        addr.set_port(port);
-                    }
-                    by_pref.push(ResolvedAddress {
-                        name: mx_host.to_string(),
-                        addr: addr.into(),
-                        is_secure: false,
-                    });
-                    continue;
-                }
-
-                match ip_lookup(mx_host, None, strategy).await {
-                    Err(err) => {
-                        tracing::error!("failed to resolve {mx_host}: {err:#}");
-                        continue;
-                    }
-                    Ok((result, _expires)) => {
-                        for addr in result.addrs.iter() {
-                            let mut addr: HostOrSocketAddress = (*addr).into();
-                            if let Some(port) = opt_port {
-                                addr.set_port(port);
-                            }
-                            by_pref.push(ResolvedAddress {
-                                name: mx_host.to_string(),
-                                addr,
-                                is_secure: result.secure,
-                            });
-                        }
-                    }
-                }
-            }
-
-            // Randomize the list of addresses within this preference
-            // level. This probablistically "load balances" outgoing
-            // traffic across MX hosts with equal preference value.
-            let mut rng = rand::thread_rng();
-            by_pref.shuffle(&mut rng);
-            result.append(&mut by_pref);
-        }
-        ResolvedMxAddresses::Addresses(result)
-    }
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub enum ResolvedMxAddresses {
-    NullMx,
-    /// The list of addresses to which to connect, expressed
-    /// in LIFO order
-    Addresses(Vec<ResolvedAddress>),
-}
-
-struct ByPreference {
-    hosts: Vec<String>,
-    pref: u16,
-    is_secure: bool,
-    is_mx: bool,
-}
-
-async fn lookup_mx_record(domain_name: &Name) -> anyhow::Result<(Vec<ByPreference>, Instant)> {
-    let mx_lookup = timeout(get_mx_timeout(), async {
-        let _permit = MX_CONCURRENCY_SEMA.acquire().await;
-        RESOLVER
-            .load()
-            .resolve(domain_name.clone(), RecordType::MX)
-            .await
-    })
-    .await??;
-    let mx_records = mx_lookup.records;
-
-    if mx_records.is_empty() {
-        if mx_lookup.nxdomain {
-            anyhow::bail!("NXDOMAIN");
-        }
-
-        // No MX records: the domain's own A/AAAA records act as the implicit
-        // MX. This implicit MX is secure exactly when the MX NODATA response
-        // was securely (DNSSEC) resolved, which is common for signed domains
-        // that publish no MX (e.g. many `.br` domains).
-        return Ok((
-            vec![ByPreference {
-                hosts: vec![domain_name.to_lowercase().to_ascii()],
-                pref: 1,
-                is_secure: mx_lookup.secure,
-                is_mx: false,
-            }],
-            mx_lookup.expires,
-        ));
-    }
-
-    let mut records: Vec<ByPreference> = Vec::with_capacity(mx_records.len());
-
-    for mx_record in mx_records {
-        if let RData::MX(mx) = mx_record {
-            let pref = mx.preference;
-            let host = mx.exchange.to_lowercase().to_string();
-
-            if let Some(record) = records.iter_mut().find(|r| r.pref == pref) {
-                record.hosts.push(host);
-            } else {
-                records.push(ByPreference {
-                    hosts: vec![host],
-                    pref,
-                    is_secure: mx_lookup.secure,
-                    is_mx: true,
-                });
-            }
-        }
-    }
-
-    // Sort by preference
-    records.sort_unstable_by(|a, b| a.pref.cmp(&b.pref));
-
-    // Sort the hosts at each preference level to produce the
-    // overall ordered list of hosts for this site
-    for mx in &mut records {
-        mx.hosts.sort();
-    }
-
-    Ok((records, mx_lookup.expires))
-}
-
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Default, Hash)]
 #[repr(u8)]
 pub enum IpLookupStrategy {
@@ -1071,84 +714,6 @@ pub async fn ipv6_lookup(
             .await;
     }
     Ok((result, expires))
-}
-
-/// Given a list of host names, produce a pseudo-regex style alternation list
-/// of the different elements of the hostnames.
-/// The goal is to produce a more compact representation of the name list
-/// with the common components factored out.
-fn factor_names<S: AsRef<str>>(name_strings: &[S]) -> String {
-    let mut max_element_count = 0;
-
-    let mut names = vec![];
-
-    for name in name_strings {
-        let (name, opt_port) = match has_colon_port(name.as_ref()) {
-            Some((name, port)) => (name, Some(port)),
-            None => (name.as_ref(), None),
-        };
-        if let Ok(name) = fully_qualify(name) {
-            names.push((name.to_lowercase(), opt_port));
-        }
-    }
-
-    let mut elements: Vec<Vec<&str>> = vec![];
-
-    let mut split_names = vec![];
-    for (name, opt_port) in names {
-        let mut fields: Vec<_> = name
-            .iter()
-            .map(|s| String::from_utf8_lossy(s).to_string())
-            .collect();
-        if let Some(port) = opt_port {
-            fields.last_mut().map(|s| {
-                s.push_str(&format!(":{port}"));
-            });
-        }
-        fields.reverse();
-        max_element_count = max_element_count.max(fields.len());
-        split_names.push(fields);
-    }
-
-    fn add_element<'a>(elements: &mut Vec<Vec<&'a str>>, field: &'a str, i: usize) {
-        match elements.get_mut(i) {
-            Some(ele) => {
-                if !ele.contains(&field) {
-                    ele.push(field);
-                }
-            }
-            None => {
-                elements.push(vec![field]);
-            }
-        }
-    }
-
-    for fields in &split_names {
-        for (i, field) in fields.iter().enumerate() {
-            add_element(&mut elements, field, i);
-        }
-        for i in fields.len()..max_element_count {
-            add_element(&mut elements, "?", i);
-        }
-    }
-
-    let mut result = vec![];
-    for mut ele in elements {
-        let has_q = ele.contains(&"?");
-        ele.retain(|&e| e != "?");
-        let mut item_text = if ele.len() == 1 {
-            ele[0].to_string()
-        } else {
-            format!("({})", ele.join("|"))
-        };
-        if has_q {
-            item_text.push('?');
-        }
-        result.push(item_text);
-    }
-    result.reverse();
-
-    result.join(".")
 }
 
 #[cfg(test)]
@@ -1479,377 +1044,6 @@ mod test {
         );
     }
 
-    #[tokio::test]
-    async fn literal_resolve() {
-        let v4_loopback = MailExchanger::resolve("[127.0.0.1]").await.unwrap();
-        k9::snapshot!(
-            &v4_loopback,
-            r#"
-MailExchanger {
-    domain_name: "[127.0.0.1]",
-    hosts: [
-        "127.0.0.1",
-    ],
-    site_name: "127.0.0.1",
-    by_pref: {
-        1: [
-            "127.0.0.1",
-        ],
-    },
-    is_domain_literal: true,
-    is_secure: false,
-    is_mx: false,
-    expires: None,
-}
-"#
-        );
-        k9::snapshot!(
-            v4_loopback
-                .resolve_addresses(IpLookupStrategy::default())
-                .await,
-            r#"
-Addresses(
-    [
-        ResolvedAddress {
-            name: "127.0.0.1",
-            addr: 127.0.0.1,
-            is_secure: false,
-        },
-    ],
-)
-"#
-        );
-
-        let v6_loopback_non_conforming = MailExchanger::resolve("[::1]").await.unwrap();
-        k9::snapshot!(
-            &v6_loopback_non_conforming,
-            r#"
-MailExchanger {
-    domain_name: "[::1]",
-    hosts: [
-        "::1",
-    ],
-    site_name: "::1",
-    by_pref: {
-        1: [
-            "::1",
-        ],
-    },
-    is_domain_literal: true,
-    is_secure: false,
-    is_mx: false,
-    expires: None,
-}
-"#
-        );
-        k9::snapshot!(
-            v6_loopback_non_conforming
-                .resolve_addresses(IpLookupStrategy::default())
-                .await,
-            r#"
-Addresses(
-    [
-        ResolvedAddress {
-            name: "::1",
-            addr: ::1,
-            is_secure: false,
-        },
-    ],
-)
-"#
-        );
-
-        let v6_loopback = MailExchanger::resolve("[IPv6:::1]").await.unwrap();
-        k9::snapshot!(
-            &v6_loopback,
-            r#"
-MailExchanger {
-    domain_name: "[IPv6:::1]",
-    hosts: [
-        "::1",
-    ],
-    site_name: "::1",
-    by_pref: {
-        1: [
-            "::1",
-        ],
-    },
-    is_domain_literal: true,
-    is_secure: false,
-    is_mx: false,
-    expires: None,
-}
-"#
-        );
-        k9::snapshot!(
-            v6_loopback
-                .resolve_addresses(IpLookupStrategy::default())
-                .await,
-            r#"
-Addresses(
-    [
-        ResolvedAddress {
-            name: "::1",
-            addr: ::1,
-            is_secure: false,
-        },
-    ],
-)
-"#
-        );
-    }
-
-    #[test]
-    fn name_factoring() {
-        assert_eq!(
-            factor_names(&[
-                "mta5.am0.yahoodns.net",
-                "mta6.am0.yahoodns.net",
-                "mta7.am0.yahoodns.net"
-            ]),
-            "(mta5|mta6|mta7).am0.yahoodns.net".to_string()
-        );
-
-        // Verify that the case is normalized to lowercase
-        assert_eq!(
-            factor_names(&[
-                "mta5.AM0.yahoodns.net",
-                "mta6.am0.yAHOodns.net",
-                "mta7.am0.yahoodns.net"
-            ]),
-            "(mta5|mta6|mta7).am0.yahoodns.net".to_string()
-        );
-
-        // When the names have mismatched lengths, do we produce
-        // something reasonable?
-        assert_eq!(
-            factor_names(&[
-                "gmail-smtp-in.l.google.com",
-                "alt1.gmail-smtp-in.l.google.com",
-                "alt2.gmail-smtp-in.l.google.com",
-                "alt3.gmail-smtp-in.l.google.com",
-                "alt4.gmail-smtp-in.l.google.com",
-            ]),
-            "(alt1|alt2|alt3|alt4)?.gmail-smtp-in.l.google.com".to_string()
-        );
-
-        assert_eq!(
-            factor_names(&[
-                "mta5.am0.yahoodns.net:123",
-                "mta6.am0.yahoodns.net:123",
-                "mta7.am0.yahoodns.net:123"
-            ]),
-            "(mta5|mta6|mta7).am0.yahoodns.net:123".to_string()
-        );
-        assert_eq!(
-            factor_names(&[
-                "mta5.am0.yahoodns.net:123",
-                "mta6.am0.yahoodns.net:456",
-                "mta7.am0.yahoodns.net:123"
-            ]),
-            "(mta5|mta6|mta7).am0.yahoodns.(net:123|net:456)".to_string()
-        );
-    }
-
-    /// Verify that the order is preserved and that we treat these two
-    /// examples of differently ordered sets of the same names as two
-    /// separate site name strings
-    #[test]
-    fn mx_order_name_factor() {
-        assert_eq!(
-            factor_names(&[
-                "example-com.mail.protection.outlook.com.",
-                "mx-biz.mail.am0.yahoodns.net.",
-                "mx-biz.mail.am0.yahoodns.net.",
-            ]),
-            "(example-com|mx-biz).mail.(protection|am0).(outlook|yahoodns).(com|net)".to_string()
-        );
-        assert_eq!(
-            factor_names(&[
-                "mx-biz.mail.am0.yahoodns.net.",
-                "mx-biz.mail.am0.yahoodns.net.",
-                "example-com.mail.protection.outlook.com.",
-            ]),
-            "(mx-biz|example-com).mail.(am0|protection).(yahoodns|outlook).(net|com)".to_string()
-        );
-    }
-
-    #[cfg(feature = "live-dns-tests")]
-    #[tokio::test]
-    async fn lookup_gmail_mx() {
-        let mut gmail = (*MailExchanger::resolve("gmail.com").await.unwrap()).clone();
-        gmail.expires.take();
-        k9::snapshot!(
-            &gmail,
-            r#"
-MailExchanger {
-    domain_name: "gmail.com.",
-    hosts: [
-        "gmail-smtp-in.l.google.com.",
-        "alt1.gmail-smtp-in.l.google.com.",
-        "alt2.gmail-smtp-in.l.google.com.",
-        "alt3.gmail-smtp-in.l.google.com.",
-        "alt4.gmail-smtp-in.l.google.com.",
-    ],
-    site_name: "(alt1|alt2|alt3|alt4)?.gmail-smtp-in.l.google.com",
-    by_pref: {
-        5: [
-            "gmail-smtp-in.l.google.com.",
-        ],
-        10: [
-            "alt1.gmail-smtp-in.l.google.com.",
-        ],
-        20: [
-            "alt2.gmail-smtp-in.l.google.com.",
-        ],
-        30: [
-            "alt3.gmail-smtp-in.l.google.com.",
-        ],
-        40: [
-            "alt4.gmail-smtp-in.l.google.com.",
-        ],
-    },
-    is_domain_literal: false,
-    is_secure: false,
-    is_mx: true,
-    expires: None,
-}
-"#
-        );
-
-        // This is a bad thing to have in a snapshot test really,
-        // but the whole set of live-dns-tests are already inherently
-        // unstable and flakey anyway.
-        // The main thing we expect to see here is that the list of
-        // names starts with alt4 and goes backwards through the priority
-        // order such that the last element is gmail-smtp.
-        // We expect the addresses within a given preference level to
-        // be randomized, because that is what resolve_addresses does.
-        k9::snapshot!(
-            gmail.resolve_addresses(IpLookupStrategy::default()).await,
-            r#"
-Addresses(
-    [
-        ResolvedAddress {
-            name: "alt4.gmail-smtp-in.l.google.com.",
-            addr: 2607:f8b0:4023:401::1b,
-        },
-        ResolvedAddress {
-            name: "alt4.gmail-smtp-in.l.google.com.",
-            addr: 173.194.77.27,
-        },
-        ResolvedAddress {
-            name: "alt3.gmail-smtp-in.l.google.com.",
-            addr: 2607:f8b0:4023:1::1a,
-        },
-        ResolvedAddress {
-            name: "alt3.gmail-smtp-in.l.google.com.",
-            addr: 172.253.113.26,
-        },
-        ResolvedAddress {
-            name: "alt2.gmail-smtp-in.l.google.com.",
-            addr: 2607:f8b0:4001:c1d::1b,
-        },
-        ResolvedAddress {
-            name: "alt2.gmail-smtp-in.l.google.com.",
-            addr: 74.125.126.27,
-        },
-        ResolvedAddress {
-            name: "alt1.gmail-smtp-in.l.google.com.",
-            addr: 2607:f8b0:4003:c04::1b,
-        },
-        ResolvedAddress {
-            name: "alt1.gmail-smtp-in.l.google.com.",
-            addr: 108.177.104.27,
-        },
-        ResolvedAddress {
-            name: "gmail-smtp-in.l.google.com.",
-            addr: 2607:f8b0:4023:c06::1b,
-        },
-        ResolvedAddress {
-            name: "gmail-smtp-in.l.google.com.",
-            addr: 142.251.2.26,
-        },
-    ],
-)
-"#
-        );
-    }
-
-    #[cfg(feature = "live-dns-tests")]
-    #[tokio::test]
-    async fn lookup_punycode_no_mx_only_a() {
-        let mx = MailExchanger::resolve("xn--bb-eka.at").await.unwrap();
-        assert_eq!(mx.domain_name, "xn--bb-eka.at.");
-        assert_eq!(mx.hosts[0], "xn--bb-eka.at.");
-    }
-
-    #[cfg(feature = "live-dns-tests")]
-    #[tokio::test]
-    async fn lookup_bogus_aasland() {
-        let err = MailExchanger::resolve("not-mairs.aasland.com")
-            .await
-            .unwrap_err();
-        k9::snapshot!(err, "MX lookup for not-mairs.aasland.com failed: NXDOMAIN");
-    }
-
-    // Asserts a DNSSEC-secure result, so it only holds when the default
-    // resolver validates, i.e. the unbound backend.
-    #[cfg(all(feature = "live-dns-tests", feature = "default-unbound"))]
-    #[tokio::test]
-    async fn lookup_example_com() {
-        // Has a NULL MX record
-        let mx = MailExchanger::resolve("example.com").await.unwrap();
-        k9::snapshot!(
-            mx,
-            r#"
-MailExchanger {
-    domain_name: "example.com.",
-    hosts: [
-        ".",
-    ],
-    site_name: "",
-    by_pref: {
-        0: [
-            ".",
-        ],
-    },
-    is_domain_literal: false,
-    is_secure: true,
-    is_mx: true,
-}
-"#
-        );
-    }
-
-    // Asserts a DNSSEC-secure result, so it only holds when the default
-    // resolver validates, i.e. the unbound backend.
-    #[cfg(all(feature = "live-dns-tests", feature = "default-unbound"))]
-    #[tokio::test]
-    async fn lookup_have_dane() {
-        let mx = MailExchanger::resolve("do.havedane.net").await.unwrap();
-        k9::snapshot!(
-            mx,
-            r#"
-MailExchanger {
-    domain_name: "do.havedane.net.",
-    hosts: [
-        "do.havedane.net.",
-    ],
-    site_name: "do.havedane.net",
-    by_pref: {
-        10: [
-            "do.havedane.net.",
-        ],
-    },
-    is_domain_literal: false,
-    is_secure: true,
-    is_mx: true,
-}
-"#
-        );
-    }
-
     // Requires DNSSEC-validated TLSA records, so it only holds when the default
     // resolver validates, i.e. the unbound backend.
     #[cfg(all(feature = "live-dns-tests", feature = "default-unbound"))]
@@ -1942,33 +1136,6 @@ MailExchanger {
     },
 ]
 "
-        );
-    }
-
-    #[cfg(feature = "live-dns-tests")]
-    #[tokio::test]
-    async fn mx_lookup_www_example_com() {
-        // Has no MX, should fall back to A lookup
-        let mx = MailExchanger::resolve("www.example.com").await.unwrap();
-        k9::snapshot!(
-            mx,
-            r#"
-MailExchanger {
-    domain_name: "www.example.com.",
-    hosts: [
-        "www.example.com.",
-    ],
-    site_name: "www.example.com",
-    by_pref: {
-        1: [
-            "www.example.com.",
-        ],
-    },
-    is_domain_literal: false,
-    is_secure: false,
-    is_mx: false,
-}
-"#
         );
     }
