+++ crates/kumo-api-types/src/shaping.rs
@@ -6,7 +6,8 @@ use config::any_err;
 #[cfg(feature = "lua")]
 use config::serialize_options;
 #[cfg(feature = "lua")]
-use dns_resolver::{fully_qualify, MailExchanger};
+use dns_resolver::fully_qualify;
+use mailexchanger::MailExchanger;
 #[cfg(feature = "lua")]
 use kumo_log_types::JsonLogRecord;
 #[cfg(feature = "lua")]
