+++ crates/mailexchanger/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "dns-resolver"
+name = "mailexchanger"
 version = "0.1.0"
 edition = "2021"
 
@@ -7,27 +7,24 @@ edition = "2021"
 # Enable this feature to run real DNS lookups.
 # Resulting tests may be flaky and, if the upstream
 # has changed their records, may be wrong.
-live-dns-tests = []
-# Make the default configuration use unbound
-default-unbound = ["unbound"]
-#default = ["default-unbound"]
-unbound = ["dep:libunbound"]
+live-dns-tests = ["dns-resolver/live-dns-tests"]
+default-unbound = ["dns-resolver/default-unbound"]
 
 [dependencies]
 anyhow = {workspace=true}
-arc-swap = {workspace=true}
-async-trait = {workspace=true}
+dns-resolver = {path="../dns-resolver"}
 kumo-address = {path="../kumo-address"}
 kumo-log-types = {path="../kumo-log-types"}
-libunbound = {workspace=true, optional=true}
+kumo-prometheus = {path="../kumo-prometheus"}
 linkme.workspace = true
 lruttl = {path="../lruttl"}
+mta-sts = {path="../mta-sts"}
+rand = {workspace=true}
 serde = {workspace=true}
-thiserror = {workspace=true}
-tokio = {workspace=true, features=["macros"]}
+tokio = {workspace=true, features=["macros", "sync", "time"]}
 tracing = {workspace=true}
-hickory-proto = {workspace=true}
 hickory-resolver = {workspace=true}
 
 [dev-dependencies]
 k9 = {workspace=true}
+tokio = {workspace=true, features=["macros"]}
