+++ crates/mod-dns-resolver/Cargo.toml
@@ -8,6 +8,7 @@ anyhow = {workspace=true}
 async-trait.workspace = true
 config = {path="../config"}
 dns-resolver = {path="../dns-resolver", features=["unbound"]}
+mailexchanger = {path="../mailexchanger"}
 duration-serde = {path="../duration-serde"}
 kumo-address = {path="../kumo-address"}
 libunbound = {workspace=true}
